requirements.txt
annotated-types==0.7.0
anyio==4.9.0
babel==2.17.0
certifi==2025.1.31
distro==1.9.0
docxcompose==1.4.0
dotenv==0.9.9
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
lxml==5.3.1
MarkupSafe==3.0.2
openai==1.70.0
pydantic==2.11.1
pydantic_core==2.33.0
python-docx==1.1.2
python-dotenv==1.1.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
